Correction faille de sécurité dans module notifier. Le nombre de fenêtres de notification est désormais limité à 5 simultanément, cela empêche d'ouvrir assez de fenêtres pour planter l'interface graphique.
git-svn-id: svn://localhost/qrezix/trunk@1834 30024bfc-89fa-0310-95ff-ee513f6fc940

+++ main/main.cpp
@@ -60,7 +60,7 @@ void fatalHandler( int signum )
 		qDebug( "[frame %i]: %s", i, strings[ i ] );
 	}
 	qDebug( "Received a %i signal, automatic backtrace", signum );
-	qDebug( "State of the stack: %i frames", size );
+	qDebug( "State of the stack: %i frames", (int)size );
 
 	QApplication::exit( 1 );
 }
